requirements.txt
py-clob-client>=0.29.0
requests>=2.31.0
python-dotenv>=1.0.0
web3>=6.14.0
